Added np.dtype, np.find_common_type, casting mechanism (WIP) and revamped expl/implicit casts.
As part of writing the upcasting mechanism for math ops.
- Added IMemoryBlock.Cast extension.
- Added constructor for IncorrectShapeException
- Cleanup

+++ src/NumSharp.Core/APIs/np.array_manipulation.cs
@@ -16,7 +16,7 @@ namespace NumSharp
         /// Elements that roll beyond the last position are re-introduced at the first.
         /// </summary>
         public static int roll(NDArray nd, int shift, int axis = -1)
-            => nd.roll(shift, axis);
+            => (int) nd.roll(shift, axis);
 
         /// <summary>
         /// Return a contiguous flattened array.
